requirements.txt
alabaster==0.7.12
amqp==2.5.2
asgiref==3.2.3
Babel==2.7.0
beautifulsoup4==4.7.1
billiard==3.6.2.0
celery==4.4.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==2.2.10
django-appconf==1.0.3
django-axes==5.0.20
django-celery-beat==1.6.0
django-ckeditor==5.8.0
django-debug-toolbar==2.1
django-filter==2.2.0
django-htmlmin==0.10.0
django-ipware==2.1.0
django-js-asset==1.2.2
django-rest-swagger==2.2.0
django-robots==3.1.0
django-timezone-field==3.1
django-widget-tweaks==1.4.3
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
imagesize==1.1.0
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.10.3
Khayyam==3.0.17
kiwisolver==1.1.0
kombu==4.6.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.2
openapi-codec==1.3.2
packaging==19.2
Pillow==6.2.0
psycopg2==2.7.7
psycopg2-binary==2.8.4
Pygments==2.4.2
pyparsing==2.4.2
python-crontab==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
simplejson==3.17.0
six==1.12.0
snowballstemmer==2.0.0
soupsieve==1.8
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
webencodings==0.5.1
zipp==2.1.0
